Extract game load progress into modules under the size ratchet
Keep AppLoadingScreen as the mascot splash, stream GET /api/games
through a dedicated fetch helper, and abort in-flight downloads.
