+++ apps/web/src/UnpublishedPlayView.tsx
@@ -1,8 +1,9 @@
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { AppLoadingScreen } from './AppLoadingScreen.js';
-import { fetchPublishedGame, type FetchProgress, type GameFetchError } from './catalog.js';
+import { GameLoadScreen } from './GameLoadScreen.js';
+import { type GameFetchError } from './catalog.js';
 import { GameTheater } from './GameTheater.js';
+import { usePublishedGameFetch } from './usePublishedGameFetch.js';
 
 type UnpublishedPlayViewProps = {
   slug: string;
@@ -25,8 +26,7 @@ type UnpublishedPlayViewProps = {
  */
 export function UnpublishedPlayView({ slug, onExit, onTitle }: UnpublishedPlayViewProps) {
   const { t } = useTranslation();
-  const [game, setGame] = useState<{ title: string; html: string } | null>(null);
-  const [loadProgress, setLoadProgress] = useState<FetchProgress>({ loaded: 0, total: null });
+  const { game, progress, error: fetchError } = usePublishedGameFetch(slug);
   const [error, setError] = useState<string | null>(null);
 
   useEffect(() => {
@@ -38,36 +38,16 @@ export function UnpublishedPlayView({ slug, onExit, onTitle }: UnpublishedPlayVi
   }, [onTitle]);
 
   useEffect(() => {
-    let cancelled = false;
-    const abort = new AbortController();
-    setGame(null);
-    setLoadProgress({ loaded: 0, total: null });
-    setError(null);
+    if (!fetchError) {
+      setError(null);
+      return;
+    }
+    const status = (fetchError as GameFetchError).status;
+    // 404/409: not shared / not ready / unknown. Anything else is a glitch — don't
+    // tell the owner their draft vanished when the request just failed.
+    setError(status === 404 || status === 409 ? t('draft.notFound') : t('draft.error'));
+  }, [fetchError, t]);
 
-    fetchPublishedGame(slug, {
-      signal: abort.signal,
-      onProgress: (progress) => {
-        if (!cancelled) setLoadProgress(progress);
-      },
-    })
-      .then((result) => {
-        if (!cancelled) setGame({ title: result.title, html: result.html });
-      })
-      .catch((err: unknown) => {
-        if (cancelled || (err instanceof Error && err.name === 'AbortError')) return;
-        const status = (err as GameFetchError).status;
-        // 404/409: not shared / not ready / unknown. Anything else is a glitch — don't
-        // tell the owner their draft vanished when the request just failed.
-        setError(status === 404 || status === 409 ? t('draft.notFound') : t('draft.error'));
-      });
-
-    return () => {
-      cancelled = true;
-      abort.abort();
-    };
-  }, [slug, t]);
-
-  // Lock the page for loading and play — both are full-viewport overlays.
   useEffect(() => {
     if (error) return;
     document.body.classList.add('player-open');
@@ -96,8 +76,7 @@ export function UnpublishedPlayView({ slug, onExit, onTitle }: UnpublishedPlayVi
   }
 
   if (!game) {
-    // Same overlay as published play, with download progress.
-    return <AppLoadingScreen onExit={onExit} progress={loadProgress} />;
+    return <GameLoadScreen onExit={onExit} progress={progress} />;
   }
 
   return (
